Fix issue for none option
There was an issue where the option was always false or the initial none as the value
was not reaching the correct comparision as the original value of the form was returned as string
from a call of the DB, but is rendered as an int on the usage of get_field.

With this change we make sure we are comparing always integer values.